refactor(adapters): extract extractHares to shared module + fix #1201 #1177
- Move extractHares, DEFAULT_HARE_PATTERNS, and PHONE_TRAILING_RE from
  google-calendar/adapter.ts to a new hare-extraction.ts module so meetup
  and phoenixhhh stop reaching across adapter boundaries. Tests moved
  verbatim to hare-extraction.test.ts. Lives outside utils.ts to avoid an
  import cycle through audit-checks.
- Add a dedicated FFMH3 source row for JEM category/8. The existing cat/3
  source is the multi-kennel aggregator (1827 FH3 + 61 FFMH3 + 27 SHITS +
  15 DOM + 1 Bike historical entries) — swapping its URL would lose the
  four sibling kennels, so cat/8 is added as a second source instead.
- Update EH3 Edmonton seed URL from wp-json pages/423 to /pages/437. The
  adapter has a hardcoded PAGE_KENNEL_MAP and never reads source.url at
  runtime, so this is metadata-only for admin UI / audit display.

Closes #1201
Closes #1177
